Creating interface to interact with announcments table (created locally). Moved executeQuery() to connect.js and export since it will be used across files.
