+++ api/app.js
@@ -6,6 +6,7 @@ var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
+var announcementsRouter = require('./routes/announcements');
 var tempRouter = require('./routes/temp');
 
 var app = express();
@@ -22,6 +23,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/announcements', announcementsRouter);
 app.use('/end', tempRouter)
 
 // catch 404 and forward to error handler
