+++ api/db/connect.js
@@ -14,6 +14,30 @@ const pool = new Pool({
   connectionTimeoutMillis: 2000,
 });
 
+function executeQuery(query, res) {
+  pool.connect((err, client, release) => {
+    if (err) {
+      return console.error('Error acquiring client', err.stack);
+    }
+    client.query(query, (err, result) => {
+      release();
+      if (err) {
+        return console.error('Error executing query', err.stack);
+      }
+      console.log(result.rows);
+      res.status(200).send(result.rows);
+    })
+  })
+  // Cannot call pool.end more than once. Would have to create an entirely new pool...
+  // pool.end((err) => {
+  //   console.log('client has disconnected');
+  //   // res.status(200).send('client has disconnected');
+  //   if (err) {
+  //     console.log('error during disconnection', err.stack);
+  //   }
+  // })
+}
+
 function endConnection(req, res, next) {
   console.log('calling endConnection');
   
@@ -28,5 +52,6 @@ function endConnection(req, res, next) {
 
 module.exports = {
   pool,
+  executeQuery,
   endConnection
 };
